requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
distlib==0.3.6
filelock==3.8.0
idna==3.3
platformdirs==2.5.2
psycopg2==2.9.3
python-dotenv==0.21.0
pytube==12.1.0
requests==2.28.1
urllib3==1.26.12
virtualenv==20.16.4
